Moved constants to DatabaseConstants class
