+++ collect_app/src/main/java/org/odk/collect/android/database/DatabaseConstants.java
@@ -2,8 +2,13 @@ package org.odk.collect.android.database;
 
 public class DatabaseConstants {
 
+    public static final String FORMS_DATABASE_NAME = "forms.db";
     public static final String FORMS_TABLE_NAME = "forms";
     public static final int FORMS_DATABASE_VERSION = 10;
+
+    public static final String INSTANCES_DATABASE_NAME = "instances.db";
+    public static final String INSTANCES_TABLE_NAME = "instances";
+    public static final int INSTANCES_DATABASE_VERSION = 6;
 
     private DatabaseConstants() {
