+++ collect_app/src/main/java/org/odk/collect/android/provider/InstanceProvider.java
@@ -45,7 +45,7 @@ import java.util.Locale;
 
 import timber.log.Timber;
 
-import static org.odk.collect.android.database.InstanceDatabaseMigrator.INSTANCES_TABLE_NAME;
+import static org.odk.collect.android.database.DatabaseConstants.INSTANCES_TABLE_NAME;
 import static org.odk.collect.android.utilities.PermissionUtils.areStoragePermissionsGranted;
 
 public class InstanceProvider extends ContentProvider {
